+++ Assets/Scripts/PlayerScripts/InputHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2d4eb40bbfdd06345888390b13411da1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
