+++ Assets/Scripts/PlayerScripts/PlayerMover.cs
@@ -1,0 +1,34 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+
+[RequireComponent(typeof(InputHandler))]
+public class PlayerMover : MonoBehaviour
+{
+    private InputHandler _input;
+
+    [SerializeField]
+    private float moveSpeed;
+
+    [SerializeField]
+    private Camera Camera;
+
+    void Awake()
+    {
+        _input = GetComponent<InputHandler>();
+    }
+
+    void Update()
+    {
+        var targetVector = new Vector3(_input.InputVector.x, 0, _input.InputVector.y);
+        Move(targetVector);
+    }
+    private void Move(Vector3 targetVector)
+    {
+        var speed = moveSpeed * Time.deltaTime;
+        targetVector = Quaternion.Euler(0, Camera.gameObject.transform.rotation.eulerAngles.y, 0) * targetVector;
+        var targetPosition = transform.position + targetVector * speed;
+        transform.position = targetPosition;
+    }
+}
