+++ Assets/Scripts/PlayerScripts/PlayerMover.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0df9d49b24ad267429eb23f3be405447
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
